暂存

+++ Taf.Core.Net.Utility/Pagging/IHasTotalCount.cs
@@ -23,5 +23,5 @@ public interface IHasTotalCount
     /// <summary>
     /// Total count of Items.
     /// </summary>
-    long TotalCount{ get; set; }
+    int TotalCount{ get; set; }
 }
